Improved generate
