+++ src/index.ts
@@ -1,5 +1,226 @@
-import * as AllIcons from './Icons';
-
 export { iconList } from './iconList';
-export * from './Icons';
-export type Icons = keyof typeof AllIcons;
+
+export type IconProps = 'Photo' | 'Component' | 'Grid' | 'Outline' | 'PhotoDrag' | 'GridAlt' | 'Search' | 'Zoom' | 'ZoomOut' | 'ZoomReset' | 'Eye' | 'EyeClose' | 'Lightning' | 'LightningOff' | 'Contrast' | 'SwitchAlt' | 'Mirror' | 'Grow' | 'PaintBrush' | 'Ruler' | 'Stop' | 'Camera' | 'Video' | 'Speaker' | 'Play' | 'PlayBack' | 'PlayNext' | 'Rewind' | 'FastForward' | 'StopAlt' | 'SideBySide' | 'Stacked' | 'Sun' | 'Moon' | 'Book' | 'Document' | 'Copy' | 'Category' | 'Folder' | 'Print' | 'GraphLine' | 'Calendar' | 'GraphBar' | 'Menu' | 'MenuReverse' | 'Filter' | 'DocChart' | 'DocList' | 'Drag' | 'Markup' | 'Bold' | 'Italic' | 'PaperClip' | 'ListOrdered' | 'ListUnordered' | 'Paragraph' | 'Markdown' | 'Repo' | 'Commit' | 'Branch' | 'PullRequest' | 'Merge' | 'Apple' | 'Linux' | 'Ubuntu' | 'Windows' | 'Chrome' | 'Storybook' | 'AzureDevOps' | 'Bitbucket' | 'Chromatic' | 'ComponentDriven' | 'Discord' | 'Facebook' | 'Figma' | 'GDrive' | 'Github' | 'Gitlab' | 'Google' | 'Graphql' | 'Medium' | 'Redux' | 'Twitter' | 'Youtube' | 'VSCode' | 'Linkedin' | 'Browser' | 'Tablet' | 'Mobile' | 'Watch' | 'Sidebar' | 'SidebarAlt' | 'SidebarAltToggle' | 'SidebarToggle' | 'BottomBar' | 'BottomBarToggle' | 'CPU' | 'Database' | 'Memory' | 'Structure' | 'Box' | 'Power' | 'Edit' | 'Cog' | 'Nut' | 'Wrench' | 'Ellipsis' | 'Wand' | 'Check' | 'Form' | 'BatchDeny' | 'BatchAccept' | 'Controls' | 'Plus' | 'CloseAlt' | 'Cross' | 'Trash' | 'PinAlt' | 'Unpin' | 'Add' | 'Subtract' | 'Close' | 'Delete' | 'Passed' | 'Changed' | 'Failed' | 'Clear' | 'Comment' | 'CommentAdd' | 'RequestChange' | 'Comments' | 'Chat' | 'Lock' | 'Unlock' | 'Key' | 'Outbox' | 'Credit' | 'Button' | 'Type' | 'PointerDefault' | 'PointerHand' | 'Command' | 'Info' | 'Question' | 'Support' | 'Alert' | 'AlertAlt' | 'Email' | 'Phone' | 'Link' | 'LinkBroken' | 'Bell' | 'RSS' | 'ShareAlt' | 'Share' | 'JumpTo' | 'CircleHollow' | 'Circle' | 'BookmarkHollow' | 'Bookmark' | 'Diamond' | 'HeartHollow' | 'Heart' | 'StarHollow' | 'Star' | 'Certificate' | 'Verified' | 'ThumbsUp' | 'Shield' | 'Basket' | 'Beaker' | 'Hourglass' | 'Flag' | 'CloudHollow' | 'Cloud' | 'Sticker' | 'ChevronUp' | 'ChevronDown' | 'ChevronLeft' | 'ChevronRight' | 'ArrowUp' | 'ArrowDown' | 'ArrowLeft' | 'ArrowRight' | 'ArrowSolidUp' | 'ArrowSolidDown' | 'ArrowSolidLeft' | 'ArrowSolidRight' | 'ExpandAlt' | 'Collapse' | 'Expand' | 'Unfold' | 'Transfer' | 'Redirect' | 'Undo' | 'Reply' | 'Sync' | 'Upload' | 'Download' | 'Back' | 'Proceed' | 'Refresh' | 'Globe' | 'Compass' | 'Location' | 'Pin' | 'Time' | 'Dashboard' | 'Timer' | 'Home' | 'Admin' | 'Direction' | 'User' | 'UserAlt' | 'UserAdd' | 'Users' | 'Profile' | 'FaceHappy' | 'FaceNeutral' | 'FaceSad' | 'Accessibility' | 'AccessibilityAlt';
+
+export { Photo } from './icons/Photo';
+export { Component } from './icons/Component';
+export { Grid } from './icons/Grid';
+export { Outline } from './icons/Outline';
+export { PhotoDrag } from './icons/PhotoDrag';
+export { GridAlt } from './icons/GridAlt';
+export { Search } from './icons/Search';
+export { Zoom } from './icons/Zoom';
+export { ZoomOut } from './icons/ZoomOut';
+export { ZoomReset } from './icons/ZoomReset';
+export { Eye } from './icons/Eye';
+export { EyeClose } from './icons/EyeClose';
+export { Lightning } from './icons/Lightning';
+export { LightningOff } from './icons/LightningOff';
+export { Contrast } from './icons/Contrast';
+export { SwitchAlt } from './icons/SwitchAlt';
+export { Mirror } from './icons/Mirror';
+export { Grow } from './icons/Grow';
+export { PaintBrush } from './icons/PaintBrush';
+export { Ruler } from './icons/Ruler';
+export { Stop } from './icons/Stop';
+export { Camera } from './icons/Camera';
+export { Video } from './icons/Video';
+export { Speaker } from './icons/Speaker';
+export { Play } from './icons/Play';
+export { PlayBack } from './icons/PlayBack';
+export { PlayNext } from './icons/PlayNext';
+export { Rewind } from './icons/Rewind';
+export { FastForward } from './icons/FastForward';
+export { StopAlt } from './icons/StopAlt';
+export { SideBySide } from './icons/SideBySide';
+export { Stacked } from './icons/Stacked';
+export { Sun } from './icons/Sun';
+export { Moon } from './icons/Moon';
+export { Book } from './icons/Book';
+export { Document } from './icons/Document';
+export { Copy } from './icons/Copy';
+export { Category } from './icons/Category';
+export { Folder } from './icons/Folder';
+export { Print } from './icons/Print';
+export { GraphLine } from './icons/GraphLine';
+export { Calendar } from './icons/Calendar';
+export { GraphBar } from './icons/GraphBar';
+export { Menu } from './icons/Menu';
+export { MenuReverse } from './icons/MenuReverse';
+export { Filter } from './icons/Filter';
+export { DocChart } from './icons/DocChart';
+export { DocList } from './icons/DocList';
+export { Drag } from './icons/Drag';
+export { Markup } from './icons/Markup';
+export { Bold } from './icons/Bold';
+export { Italic } from './icons/Italic';
+export { PaperClip } from './icons/PaperClip';
+export { ListOrdered } from './icons/ListOrdered';
+export { ListUnordered } from './icons/ListUnordered';
+export { Paragraph } from './icons/Paragraph';
+export { Markdown } from './icons/Markdown';
+export { Repo } from './icons/Repo';
+export { Commit } from './icons/Commit';
+export { Branch } from './icons/Branch';
+export { PullRequest } from './icons/PullRequest';
+export { Merge } from './icons/Merge';
+export { Apple } from './icons/Apple';
+export { Linux } from './icons/Linux';
+export { Ubuntu } from './icons/Ubuntu';
+export { Windows } from './icons/Windows';
+export { Chrome } from './icons/Chrome';
+export { Storybook } from './icons/Storybook';
+export { AzureDevOps } from './icons/AzureDevOps';
+export { Bitbucket } from './icons/Bitbucket';
+export { Chromatic } from './icons/Chromatic';
+export { ComponentDriven } from './icons/ComponentDriven';
+export { Discord } from './icons/Discord';
+export { Facebook } from './icons/Facebook';
+export { Figma } from './icons/Figma';
+export { GDrive } from './icons/GDrive';
+export { Github } from './icons/Github';
+export { Gitlab } from './icons/Gitlab';
+export { Google } from './icons/Google';
+export { Graphql } from './icons/Graphql';
+export { Medium } from './icons/Medium';
+export { Redux } from './icons/Redux';
+export { Twitter } from './icons/Twitter';
+export { Youtube } from './icons/Youtube';
+export { VSCode } from './icons/VSCode';
+export { Linkedin } from './icons/Linkedin';
+export { Browser } from './icons/Browser';
+export { Tablet } from './icons/Tablet';
+export { Mobile } from './icons/Mobile';
+export { Watch } from './icons/Watch';
+export { Sidebar } from './icons/Sidebar';
+export { SidebarAlt } from './icons/SidebarAlt';
+export { SidebarAltToggle } from './icons/SidebarAltToggle';
+export { SidebarToggle } from './icons/SidebarToggle';
+export { BottomBar } from './icons/BottomBar';
+export { BottomBarToggle } from './icons/BottomBarToggle';
+export { CPU } from './icons/CPU';
+export { Database } from './icons/Database';
+export { Memory } from './icons/Memory';
+export { Structure } from './icons/Structure';
+export { Box } from './icons/Box';
+export { Power } from './icons/Power';
+export { Edit } from './icons/Edit';
+export { Cog } from './icons/Cog';
+export { Nut } from './icons/Nut';
+export { Wrench } from './icons/Wrench';
+export { Ellipsis } from './icons/Ellipsis';
+export { Wand } from './icons/Wand';
+export { Check } from './icons/Check';
+export { Form } from './icons/Form';
+export { BatchDeny } from './icons/BatchDeny';
+export { BatchAccept } from './icons/BatchAccept';
+export { Controls } from './icons/Controls';
+export { Plus } from './icons/Plus';
+export { CloseAlt } from './icons/CloseAlt';
+export { Cross } from './icons/Cross';
+export { Trash } from './icons/Trash';
+export { PinAlt } from './icons/PinAlt';
+export { Unpin } from './icons/Unpin';
+export { Add } from './icons/Add';
+export { Subtract } from './icons/Subtract';
+export { Close } from './icons/Close';
+export { Delete } from './icons/Delete';
+export { Passed } from './icons/Passed';
+export { Changed } from './icons/Changed';
+export { Failed } from './icons/Failed';
+export { Clear } from './icons/Clear';
+export { Comment } from './icons/Comment';
+export { CommentAdd } from './icons/CommentAdd';
+export { RequestChange } from './icons/RequestChange';
+export { Comments } from './icons/Comments';
+export { Chat } from './icons/Chat';
+export { Lock } from './icons/Lock';
+export { Unlock } from './icons/Unlock';
+export { Key } from './icons/Key';
+export { Outbox } from './icons/Outbox';
+export { Credit } from './icons/Credit';
+export { Button } from './icons/Button';
+export { Type } from './icons/Type';
+export { PointerDefault } from './icons/PointerDefault';
+export { PointerHand } from './icons/PointerHand';
+export { Command } from './icons/Command';
+export { Info } from './icons/Info';
+export { Question } from './icons/Question';
+export { Support } from './icons/Support';
+export { Alert } from './icons/Alert';
+export { AlertAlt } from './icons/AlertAlt';
+export { Email } from './icons/Email';
+export { Phone } from './icons/Phone';
+export { Link } from './icons/Link';
+export { LinkBroken } from './icons/LinkBroken';
+export { Bell } from './icons/Bell';
+export { RSS } from './icons/RSS';
+export { ShareAlt } from './icons/ShareAlt';
+export { Share } from './icons/Share';
+export { JumpTo } from './icons/JumpTo';
+export { CircleHollow } from './icons/CircleHollow';
+export { Circle } from './icons/Circle';
+export { BookmarkHollow } from './icons/BookmarkHollow';
+export { Bookmark } from './icons/Bookmark';
+export { Diamond } from './icons/Diamond';
+export { HeartHollow } from './icons/HeartHollow';
+export { Heart } from './icons/Heart';
+export { StarHollow } from './icons/StarHollow';
+export { Star } from './icons/Star';
+export { Certificate } from './icons/Certificate';
+export { Verified } from './icons/Verified';
+export { ThumbsUp } from './icons/ThumbsUp';
+export { Shield } from './icons/Shield';
+export { Basket } from './icons/Basket';
+export { Beaker } from './icons/Beaker';
+export { Hourglass } from './icons/Hourglass';
+export { Flag } from './icons/Flag';
+export { CloudHollow } from './icons/CloudHollow';
+export { Cloud } from './icons/Cloud';
+export { Sticker } from './icons/Sticker';
+export { ChevronUp } from './icons/ChevronUp';
+export { ChevronDown } from './icons/ChevronDown';
+export { ChevronLeft } from './icons/ChevronLeft';
+export { ChevronRight } from './icons/ChevronRight';
+export { ArrowUp } from './icons/ArrowUp';
+export { ArrowDown } from './icons/ArrowDown';
+export { ArrowLeft } from './icons/ArrowLeft';
+export { ArrowRight } from './icons/ArrowRight';
+export { ArrowSolidUp } from './icons/ArrowSolidUp';
+export { ArrowSolidDown } from './icons/ArrowSolidDown';
+export { ArrowSolidLeft } from './icons/ArrowSolidLeft';
+export { ArrowSolidRight } from './icons/ArrowSolidRight';
+export { ExpandAlt } from './icons/ExpandAlt';
+export { Collapse } from './icons/Collapse';
+export { Expand } from './icons/Expand';
+export { Unfold } from './icons/Unfold';
+export { Transfer } from './icons/Transfer';
+export { Redirect } from './icons/Redirect';
+export { Undo } from './icons/Undo';
+export { Reply } from './icons/Reply';
+export { Sync } from './icons/Sync';
+export { Upload } from './icons/Upload';
+export { Download } from './icons/Download';
+export { Back } from './icons/Back';
+export { Proceed } from './icons/Proceed';
+export { Refresh } from './icons/Refresh';
+export { Globe } from './icons/Globe';
+export { Compass } from './icons/Compass';
+export { Location } from './icons/Location';
+export { Pin } from './icons/Pin';
+export { Time } from './icons/Time';
+export { Dashboard } from './icons/Dashboard';
+export { Timer } from './icons/Timer';
+export { Home } from './icons/Home';
+export { Admin } from './icons/Admin';
+export { Direction } from './icons/Direction';
+export { User } from './icons/User';
+export { UserAlt } from './icons/UserAlt';
+export { UserAdd } from './icons/UserAdd';
+export { Users } from './icons/Users';
+export { Profile } from './icons/Profile';
+export { FaceHappy } from './icons/FaceHappy';
+export { FaceNeutral } from './icons/FaceNeutral';
+export { FaceSad } from './icons/FaceSad';
+export { Accessibility } from './icons/Accessibility';
+export { AccessibilityAlt } from './icons/AccessibilityAlt';
